aviary/wrenformer/utils.py -> examples/wrenformer/mat_bench/utils.py
move print_walltime decorator to aviary/utils.py
