+++ examples/wrenformer/mat_bench/make_plots.py
@@ -13,7 +13,6 @@ from matbench.metadata import mbv01_metadata as matbench_metadata
 from pymatviz.utils import add_identity_line
 from sklearn.metrics import r2_score, roc_auc_score
 
-from aviary.wrenformer.utils import recursive_dict_merge
 from examples.wrenformer.mat_bench import DATA_PATHS
 from examples.wrenformer.mat_bench.plotting_functions import (
     dataset_labels_html,
@@ -21,6 +20,7 @@ from examples.wrenformer.mat_bench.plotting_functions import (
     plot_leaderboard,
     scale_errors,
 )
+from examples.wrenformer.mat_bench.utils import recursive_dict_merge
 
 __author__ = "Janosh Riebesell"
 __date__ = "2022-04-25"
